requirements.txt
torch==2.7.1+cu118
torchvision==0.22.1+cu118
-f https://download.pytorch.org/whl/cu118/torch_stable.html
timm
torchmetrics
transformers
albumentations
numpy==1.25
pandas
Pillow
scikit-learn
hf_xet
matplotlib
seaborn
